+++ CHANGELOG.md
@@ -18,6 +18,7 @@ Changes to this project are documented in this file. More detail (including info
 - AUTOTOOL-3985: Improve filtering for per-app requests so that apps not in the declaration are not considered in the diffs.
 - AUTOTOOL-1779: ([GitHub Issue 458](https://github.com/F5Networks/f5-appsvcs-extension/issues/458)): Service_Forwarding documentation to reflect valid options
 - AUTOTOOL-4057: Set sniDefault property to true for the first certificate by default
+- AUTOTOOL-4003: ([GitHub Issue 754](https://github.com/F5Networks/f5-appsvcs-extension/issues/754)): Can't create FQDN pool with shareNodes: true
 
 ### Changed
 - Updated to Service Discovery 1.16.0-1