+++ CHANGELOG.md
@@ -12,6 +12,8 @@ Changes to this project are documented in this file. More detail (including info
 - AUTOTOOL-4060: Certain AS3 declarations result in "declaration has duplicate values" errors
 
 ### Changed
+- Updated to Service Discovery 1.17.0-1
+  - Package updates
 
 ### Removed